feat: :art: project restructured
Restructuring of project by placing files into relevant positions.
